Prediction breakdown panel hidden by default.
Buttons placed at the bottom of the primary prediction panel which are used to toggle the visibility of the breakdown panel.

Setting added to the config panel to allow for the breakdown panel to be visible by default.

The aim is to cut down on the confusion brought on by the prediction breakdown by de-emphasizing it while still making it easily accessible.